refactor: 쿼리를 직접 작성하여 log-writer를 구현한다
