--- modules/infrastructure/log-writer/src/main/java/com/whoz_in/log_writer/infra/managed/arp/ArpLog.java
@@ -1,29 +1,0 @@
-package com.whoz_in.log_writer.infra.managed.arp;
-
-import jakarta.annotation.Nullable;
-import java.time.LocalDateTime;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-@Getter
-@EqualsAndHashCode(onlyExplicitlyIncluded = true)
-public class ArpLog {
-
-    @EqualsAndHashCode.Include
-    private String mac;
-    @EqualsAndHashCode.Include
-    private String ip;
-
-    private String device;
-    private String ssid;
-    private LocalDateTime createdAt;
-
-    public ArpLog(String mac, String ip, @Nullable String device, String ssid) {
-        this.mac = mac;
-        this.ip = ip;
-        this.device = device;
-        this.ssid = ssid;
-        this.createdAt = LocalDateTime.now();
-    }
-
-}
