+++ modules/infrastructure/log-writer/src/main/java/com/whoz_in/log_writer/infra/managed/mdns/MdnsLogParser.java
@@ -1,5 +1,6 @@
 package com.whoz_in.log_writer.infra.managed.mdns;
 
+import com.whoz_in.log_writer.infra.managed.ManagedLog;
 import java.util.Optional;
 import org.springframework.stereotype.Component;
 
@@ -11,7 +12,7 @@ public final class MdnsLogParser {
      * @param log
      * 한 줄 데이터만 들어와야 한다. 들어온 데이터중 \t가 2개가 존재하는 것만 처리한다.
      */
-    public Optional<MdnsLog> parse(String log) {
+    public Optional<ManagedLog> parse(String log) {
         String[] logParts = log.split("\t");
         if(logParts.length < 2 || logParts[0].isEmpty() || logParts[1].isEmpty()) return Optional.empty();
 
@@ -19,7 +20,7 @@ public final class MdnsLogParser {
         String ip = logParts[1];
         String deviceName = (logParts.length == 3) ? parseDeviceName(logParts[2]) : null;
 
-        return Optional.of(new MdnsLog(mac, ip, deviceName));
+        return Optional.of(new ManagedLog(mac, ip, deviceName));
     }
 
     //기기 이름 형식에 맞지 않을 때 null을 반환. 내부적으로 사용하는 메서드기 때문에 null을 반환할 수 있도록 했음
