+++ modules/infrastructure/log-writer/src/main/java/com/whoz_in/log_writer/infra/monitor/MonitorLogDAO.java
@@ -3,7 +3,6 @@ package com.whoz_in.log_writer.infra.monitor;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import java.util.Collection;
-import java.util.Set;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
